refactor(ui): split DeviceListSection and PermissionBanner out of MixerView
MixerView.swift packed three unrelated view types with the drag-reorder
math buried in the middle; the trickiest UI logic now lives in its own
file. The hand-maintained rowHeight literal moves onto AppRowView as a
constant, mirroring DeviceRowView — a layout tweak there can no longer
silently desync the scroll-frame height.
