+++ Fader/UI/PermissionBanner.swift
@@ -1,0 +1,29 @@
+import SwiftUI
+
+/// Shown when tap creation failed — almost always missing the
+/// System Audio Recording permission.
+struct PermissionBanner: View {
+    var body: some View {
+        HStack(spacing: 8) {
+            Image(systemName: "exclamationmark.triangle.fill")
+                .foregroundStyle(.yellow)
+            VStack(alignment: .leading, spacing: 2) {
+                Text("Audio access needed")
+                    .font(.system(size: 11, weight: .semibold))
+                Text("Allow Fader to record system audio in Privacy & Security.")
+                    .font(.system(size: 10))
+                    .foregroundStyle(.secondary)
+            }
+            Spacer()
+            Button("Open") {
+                let url = "x-apple.systempreferences:com.apple.preference.security?Privacy_AudioCapture"
+                if let settingsURL = URL(string: url) {
+                    NSWorkspace.shared.open(settingsURL)
+                }
+            }
+            .controlSize(.small)
+        }
+        .padding(8)
+        .background(.yellow.opacity(0.1), in: RoundedRectangle(cornerRadius: 8))
+    }
+}
